useless

+++ Zip_Files_Executables/createExe.py
@@ -2,6 +2,60 @@ import os
 import sys
 import shutil
 import subprocess
+import re
+from zipfile import ZipFile, ZIP_DEFLATED
+
+def create_zips():
+    appendage = ""
+    tag = "000000"
+    if sys.platform == "win32":
+        appendage = " (Windows)"
+        tab = "Windows"
+    elif sys.platform.startswith("linux"):
+        appendage = " (Linux)"
+        tag = "Linux"
+    elif sys.platform == "darwin":
+        appendage = " (Mac)"
+        tag = "Mac"
+
+    
+    cwd = os.getcwd()
+    files = []
+    dirs = []
+    for name in os.listdir(cwd):
+        if(os.path.isfile(name)):
+            pass
+        else:
+            dirs.append(name)
+
+    for directory in dirs:
+        print("Starting for " + directory)
+        # Copy Directory
+        copyName = str(directory + appendage)
+        shutil.copytree(directory, copyName)
+
+        # Remove executable from original
+        for file in os.listdir(directory):
+            if(tag in file):
+                if sys.platform == "darwin":
+                    subprocess.run = ("rm", "-r", os.path.join(cwd, directory, file))
+                else:
+                    os.remove(os.path.join(cwd, directory, file))
+
+        # Remove Autograder from copy
+        for file in os.listdir(copyName):
+            if(len(re.findall("lab_\\d\\d_assistant.py", file)) == 1):
+                os.remove(os.path.join(cwd, copyName, file))
+
+        # Create Zip File from copy
+        with ZipFile(copyName + ".zip", "w", ZIP_DEFLATED) as zipf:
+            for root, dirs, files in os.walk(copyName):
+                for file in files:
+                    file_path = os.path.join(root, file)
+                    zipf.write(file_path, os.path.relpath(file_path, start=copyName))
+
+        shutil.rmtree(os.path.join(cwd, copyName))
+    print("Successfully created all zip files")
 
 def testExe(pyinstaller_path):
     b_success = False
@@ -17,6 +71,15 @@ def testExe(pyinstaller_path):
             b_success = True
     except:
         pass
+
+    cwd = os.getcwd()
+    for name in os.listdir(cwd):
+        if(os.path.isfile(name)):
+            if(len(re.findall("testCompile", name)) == 1):
+                os.remove(os.path.join(cwd, name))
+        else:
+            if(len(re.findall("testCompile", name)) == 1 or name == "build" or name == "dist"):
+                shutil.rmtree(os.path.join(cwd, name))
     
     return b_success
 
@@ -59,6 +122,7 @@ def createExecutables(osName, addDataEnd):
                 appName = folder + "/" + exeName + ".app"
                 assistant = folder + "/" + folder + "_assistant.py" + addDataEnd
                 assistantImport = folder + "_assistant"
+                print(assistant, assistantImport)
                 check = folder + "/check.png" + addDataEnd
                 redX = folder + "/redX.png" + addDataEnd
                                     
@@ -90,16 +154,8 @@ def createExecutables(osName, addDataEnd):
                     print("Stderr:", result.stderr)
                     sys.exit(result.returncode)
                 os.remove(specFile)
-                shutil.rmtree(appName)
-                #result = subprocess.run(("rm", appName), capture_output=True, text=True)
-                '''if result.returncode == 0:
-                    print(exeName, "has been created!")
-                else:
-                    print("An error may have occurred")
-                    print("Stdout:", result.stdout)
-                    print("Stderr:", result.stderr)
-                    sys.exit(result.returncode)'''
-    
+                if sys.platform == "darwin":
+                    shutil.rmtree(appName)
     try:
         shutil.rmtree("build")
         print("Success! All executables are ready for use.")
@@ -138,5 +194,8 @@ def main():
         createExecutables(osName, addDataEnd)
     else:
         print(f"Unsupported OS: {sys.platform}")
+
+    create_zips()
+
 
 main()
